+++ example/lib/pages/socket/socket.dart
@@ -4,17 +4,23 @@ class SocketPage extends GetView {
   @override
   Widget build(BuildContext context) {
     return Socket(context, builder: (socket) {
-      socket.onMessage.listen((data) {
+      socket.on('join', (val) {
+        final join = socket.join(val);
+        if (join) {
+          socket.sendToRoom(val, 'socket: ${socket.hashCode} join to room');
+        }
+      });
+      socket.onMessage((data) {
         print('data: $data');
         socket.send(data);
       });
 
-      socket.onOpen.listen((ws) {
+      socket.onOpen((ws) {
         print('new socket opened');
       });
 
-      socket.onClose.listen((ws) {
-        print('socket has been closed');
+      socket.onClose((close) {
+        print('socket has closed. Reason: ${close.message}');
       });
     });
   }