+++ lib/src/socket/socket.dart
@@ -1,4 +1,5 @@
-import 'dart:async';
+import 'dart:collection';
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:get_instance/get_instance.dart';
@@ -15,19 +16,6 @@ abstract class WebSocketBase {
   void leave(String room);
 }
 
-enum SocketType {
-  message,
-  error,
-  done,
-}
-
-class SocketMessage {
-  final dynamic message;
-  final SocketType type;
-
-  SocketMessage(this.message, this.type);
-}
-
 class Close {
   final WebSocket socket;
   final String message;
@@ -36,35 +24,107 @@ class Close {
   Close(this.socket, this.message, this.reason);
 }
 
+typedef OpenSocket = void Function(WebSocket);
+
+typedef CloseSocket = void Function(Close);
+
+typedef MessageSocket = void Function(dynamic val);
+
+class SocketNotifier {
+  var _onMessages = HashSet<MessageSocket>();
+  var _onEvents = <String, MessageSocket>{};
+  var _onCloses = HashSet<CloseSocket>();
+  var _onErrors = HashSet<CloseSocket>();
+
+  void addMessages(MessageSocket socket) {
+    _onMessages.add((socket));
+  }
+
+  void addEvents(String event, MessageSocket socket) {
+    _onEvents[event] = socket;
+  }
+
+  void addCloses(CloseSocket socket) {
+    _onCloses.add(socket);
+  }
+
+  void addErrors(CloseSocket socket) {
+    _onErrors.add((socket));
+  }
+
+  void notifyData(dynamic data) {
+    for (var item in _onMessages) {
+      item(data);
+    }
+    _tryOn(data);
+  }
+
+  void notifyClose(Close err, WebSocket _ws) {
+    Get.log('Socket ${_ws.hashCode} is been disposed');
+
+    for (var item in _onCloses) {
+      item(err);
+    }
+    _onCloses = null;
+    _onErrors = null;
+    _onEvents = null;
+  }
+
+  void notifyError(Close err) {
+    // rooms.removeWhere((key, value) => value.contains(_ws));
+    for (var item in _onErrors) {
+      item(err);
+    }
+  }
+
+  void _tryOn(dynamic message) {
+    try {
+      Map msg = jsonDecode(message);
+      final event = msg['type'];
+      final data = msg['data'];
+      if (_onEvents.containsKey(event)) {
+        _onEvents[event](data);
+      }
+    } catch (err) {
+      return;
+    }
+  }
+
+  void dispose() {
+    _onMessages = null;
+    _onEvents = null;
+  }
+}
+
 class GetSocket implements WebSocketBase {
   final WebSocket _ws;
-  final _messageController = StreamController(),
-      _openController = StreamController<WebSocket>(),
-      _closeController = StreamController<Close>();
-
-  final Map<String, List<WebSocket>> rooms;
-
-  Stream _messages;
+  final Map<String, HashSet<WebSocket>> rooms;
+  SocketNotifier socketNotifier = SocketNotifier();
+  bool isDisposed = false;
 
   GetSocket(this._ws, this.rooms) {
-    _messages = _messageController.stream.asBroadcastStream();
-
-    _openController.add(_ws);
-
-    _ws.listen(_messageController.add, onError: (err) {
-      _closeController.add(Close(_ws, err.toString(), 0));
+    _ws.listen((data) {
+      socketNotifier.notifyData(data);
+    }, onError: (err) {
+      socketNotifier.notifyError(Close(_ws, err.toString(), 0));
     }, onDone: () {
-      _closeController.add(Close(_ws, 'Connection closed', 1));
+      rooms.removeWhere((key, value) => value.contains(_ws));
+      socketNotifier.notifyClose(Close(_ws, 'Connection closed', 1), _ws);
+      socketNotifier.dispose();
+      socketNotifier = null;
+      isDisposed = true;
     });
   }
 
   @override
   void send(Object message) {
+    _checkAvailable();
     _ws.add(message);
   }
 
   // TODO: Improve it
   void sendToRoom(String room, Object message) {
+    _checkAvailable();
     if (rooms.containsKey(room)) {
       rooms[room].forEach((element) {
         element.add(message);
@@ -72,8 +132,13 @@ class GetSocket implements WebSocketBase {
     }
   }
 
+  void _checkAvailable() {
+    if (isDisposed) throw 'Cannot add events to closed Socket';
+  }
+
   // TODO: Improve it
   void broadcastToRoom(String room, Object message) {
+    _checkAvailable();
     if (rooms.containsKey(room)) {
       rooms[room].forEach((element) {
         if (element != _ws) {
@@ -85,20 +150,24 @@ class GetSocket implements WebSocketBase {
 
   @override
   void emit(String event, Object data) {
-    _ws.add({event: data});
+    send({'type': event, 'data': data});
   }
 
   @override
-  void join(String room) {
+  bool join(String room) {
+    _checkAvailable();
     if (rooms.containsKey(room)) {
-      rooms[room].add(_ws);
+      return rooms[room].add(_ws);
     } else {
-      Get.log("Room $room don't exists");
+      Get.log("Room [$room] don't exists, creating it");
+      rooms[room] = HashSet();
+      return rooms[room].add(_ws);
     }
   }
 
   @override
   void leave(String room) {
+    _checkAvailable();
     if (room.contains(room)) {
       rooms[room].remove(_ws);
     } else {
@@ -106,18 +175,28 @@ class GetSocket implements WebSocketBase {
     }
   }
 
-  ///Listen messages to socket
-  Stream get onMessage => _messages;
+  void onOpen(OpenSocket fn) {
+    fn(_ws);
+  }
 
-  ///Listen socket open
-  Stream<WebSocket> get onOpen => _openController.stream;
+  void onClose(CloseSocket fn) {
+    socketNotifier.addCloses(fn);
+  }
 
-  ///Listen socket close
-  Stream<Close> get onClose => _closeController.stream;
+  void onError(CloseSocket fn) {
+    socketNotifier.addErrors(fn);
+  }
+
+  void onMessage(MessageSocket fn) {
+    socketNotifier.addMessages(fn);
+  }
+
+  void on(String event, MessageSocket message) {
+    socketNotifier.addEvents(event, message);
+  }
 
   @override
   void close([int status, String reason]) {
-    rooms.removeWhere((key, value) => value.contains(_ws));
     _ws.close(status, reason);
   }
 }